refactor: use dir modules instead of file
